Create api and dashboard modules
Moved dashboard folder under src and made it a module. Created new,
seperate api module. Tried to split up code between the two.
