+++ .gitignore
@@ -1,13 +1,9 @@
 
 # Created by https://www.gitignore.io/api/macos,python,pycharm,jupyternotebook
 
-### JupyterNotebook ###
-.ipynb_checkpoints
-*/.ipynb_checkpoints/*
+### Database ###
+config.py
 
-# Remove previous ipynb_checkpoints
-#   git rm -r .ipynb_checkpoints/
-#
 ### macOS ###
 *.DS_Store
 .AppleDouble
@@ -204,4 +200,3 @@ venv.bak/
 
 # End of https://www.gitignore.io/api/macos,python,pycharm,jupyternotebook
 *__pycache__*
-config.py
