--- dashboard/config.example.py
@@ -1,3 +1,0 @@
-DB_NAME = 'localhost'
-DB_USER = 'energycomps'
-DB_PASSWORD='test'
