+++ src/api/config.example.py
@@ -1,0 +1,6 @@
+# Make a copy of this file inside this directory, called config.py, and add the password there.
+# config.py is in the gitignore, so we don't upload the password to GitHub.
+
+DB_NAME = 'energycomps'
+DB_USER = 'energycomps'
+DB_PASSWORD='temp'
