+++ src/dashboard/static/style.css
